Refactor EET oracle structure for consistency and extensibility
Merge `EETTransformation` into a new abstract base class `EETTransformer`
(template method pattern: rules + tree-walking orchestration in the base,
`descend` abstract for DBMS-specific AST reconstruction). Move
`MySQLEETNodeFactory` and `MySQLEETTransformer` from `mysql/gen` to
`mysql/oracle` to mirror the placement of their common counterparts.
`MySQLEETTransformer` now extends `EETTransformer<MySQLExpression>`.
