+++ src/sqlancer/common/oracle/EETNodeFactory.java
@@ -3,9 +3,9 @@ package sqlancer.common.oracle;
 import sqlancer.common.ast.newast.Expression;
 
 /**
- * Factory for constructing the AST nodes needed by the {@link EETTransformation equivalent expression transformation}
- * rules. Because every DBMS has its own expression AST, the actual node construction is DBMS-specific; this interface
- * lets the (DBMS-independent) transformation rules be expressed once in terms of a small set of primitive operations.
+ * Factory for constructing the AST nodes needed by the {@link EETTransformer EET transformer's} transformation rules.
+ * Because every DBMS has its own expression AST, the actual node construction is DBMS-specific; this interface lets the
+ * (DBMS-independent) transformation rules be expressed once in terms of a small set of primitive operations.
  *
  * @param <E>
  *            the DBMS-specific expression type
